refactor : payment 비동기 적용 완료

+++ refactoring_project/src/main/java/com/paymilli/paymilli/domain/payment/dto/response/ApproveResponse.java
@@ -7,7 +7,7 @@ import lombok.Getter;
 @AllArgsConstructor
 public class ApproveResponse {
     String storeName;
-    int totalPrice;
+    long totalPrice;
     String detail;
     String refundToken;
 }
